Changed classes' methods in to static

+++ src/main/java/it/unisa/agency_formation/formazione/DAO/DocumentoDAO.java
@@ -18,17 +18,25 @@ public class DocumentoDAO {
      * @param doc
      * @throws SQLException doc!=null
      */
-    public static void doSaveDocument(Documento doc) throws SQLException {
+    public static boolean doSaveDocument(Documento doc) throws SQLException {
+        if(doc == null){
+            return false;
+        }
         Connection connection = DatabaseManager.getInstance().getConnection();
         PreparedStatement save = null;
-        String query = "SELECT INTO " + TABLE_DOCUMENTO + "(MaterialeDiFormazione, IdUtente, IdTeam) " +
+        String query = "INSERT INTO " + TABLE_DOCUMENTO + "(MaterialeDiFormazione, IdUtente, IdTeam) " +
                 "VALUES(?,?,?)";
         try {
             save = connection.prepareStatement(query);
             save.setString(1, doc.getMaterialeDiFormazione());
             save.setInt(2, doc.getIdUtente());
             save.setInt(3, doc.getIdTeam());
-            save.executeQuery();
+            int result=save.executeUpdate();
+            if(result!=-1){
+                return true;
+            }else{
+                return false;
+            }
         } finally {
             try {
                 if (save != null)
@@ -43,18 +51,27 @@ public class DocumentoDAO {
     /**
      * Questa funzionalità permette di eliminare un documento
      *
-     * @param matForm
+     * @param materialeFormazione
      * @throws SQLException
      * @pre matForm!=null
      */
-    public static void doRemoveDocument(String matForm) throws SQLException {
+    public static boolean doRemoveDocument(String materialeFormazione) throws SQLException {
+        if(materialeFormazione==null){
+            return false;
+        }
+        ResultSet result;
         Connection connection = DatabaseManager.getInstance().getConnection();
         String query = "DELETE FROM " + TABLE_DOCUMENTO + "WHERE MaterialeDiFormazione=?";
         PreparedStatement stmt = null;
         try {
             stmt = connection.prepareStatement(query);
-            stmt.setString(1, matForm);
-            stmt.executeQuery();
+            stmt.setString(1, materialeFormazione);
+            result=stmt.executeQuery();
+            if(result.next()){
+                return true;
+            }else {
+                return false;
+            }
         } finally {
             try {
                 if (stmt != null)
@@ -70,21 +87,29 @@ public class DocumentoDAO {
      * Questa funzionalità permette di modificare il materiale di formazione di un team
      *
      * @param idUtente
-     * @param ma
+     * @param materiale
      * @param idTeam
      * @throws SQLException
      * @pre idDocument != null && ma != null && idTeam!=null
      */
-    public static void updateDocument(int idUtente, String ma, int idTeam) throws SQLException {
+    public static boolean updateDocument(int idUtente, String materiale, int idTeam) throws SQLException {
+        if(idUtente <1 || materiale==null || idTeam<1){
+            return false;
+        }
         Connection connection = DatabaseManager.getInstance().getConnection();
         PreparedStatement stmt = null;
         String query = "UPDATE" + TABLE_DOCUMENTO + " SET MaterialeDiFormazione=? WHERE IdUtente=? and IdTeam=?";
         try {
             stmt = connection.prepareStatement(query);
-            stmt.setString(1, ma);
+            stmt.setString(1, materiale);
             stmt.setInt(2, idUtente);
             stmt.setInt(3, idTeam);
-            stmt.executeUpdate();
+            int result= stmt.executeUpdate();
+            if(result!=-1){
+                return true;
+            }else{
+                return false;
+            }
         } finally {
             try {
                 if (stmt != null)
@@ -105,15 +130,24 @@ public class DocumentoDAO {
      * @pre idTeam!=null
      */
     public static String doRetrieveByTeam(int idTeam) throws SQLException {
+        if(idTeam<1){
+            return null;
+        }
         Connection connection = DatabaseManager.getInstance().getConnection();
         PreparedStatement stmt = null;
         ResultSet result;
+        String path = null;
         String query = "SELECT FROM " + TABLE_DOCUMENTO + "WHERE IdTeam=?";
         try {
             stmt = connection.prepareStatement(query);
             stmt.setInt(1, idTeam);
             result = stmt.executeQuery();
-            return result.getString("MaterialeDiFormazione");
+            path = result.getString("MaterialeDiFormazione");
+            if(path!=null){
+                return path;
+            }else{
+                return null;
+            }
         } finally {
             try {
                 if (stmt != null)
